Re-organize the Starter-Kit modules, _again_
