+++ public/js/start/views/output.js
