+++ views/pages/start.handlebars
@@ -2,7 +2,7 @@
 {{setSubTitle      "Start your next web project with Pure."}}
 {{setActiveNav     "start"}}
 {{setLoadYUI       true}}
-{{setYUIBoot       "start"}}
+{{setYUIBoot       "start/app"}}
 {{setIncludePureGR true}}
 
 {{addLocalCSS "/css/start.css"}}
